revert to old indexing progress folder structure
