+++ gui/src/pages/More/IndexingProgress/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./IndexingProgress";
