--- gui/src/pages/More/progress/docs/index.tsx
@@ -1,121 +1,0 @@
-import { IndexingProgressUpdate } from "core";
-import { useContext, useEffect, useState } from "react";
-import { useDispatch } from "react-redux";
-import { IdeMessengerContext } from "../../../../context/IdeMessenger";
-import { isJetBrains } from "../../../../util";
-import { useWebviewListener } from "../../../../hooks/useWebviewListener";
-import IndexingProgressBar from "../IndexingProgressBar";
-import IndexingProgressIndicator from "../IndexingProgressIndicator";
-import IndexingProgressSubtext from "../IndexingProgressSubtext";
-import { usePostHog } from "posthog-js/react";
-import ConfirmationDialog from "../../../../components/dialogs/ConfirmationDialog";
-import {
-  setShowDialog,
-  setDialogMessage,
-} from "../../../../redux/slices/uiStateSlice";
-import IndexingProgressTitleText from "../IndexingProgressTitleText";
-import IndexingProgressErrorText from "../codebase/IndexingProgressErrorText";
-
-function DocsIndexingProgressItem() {
-  const ideMessenger = useContext(IdeMessengerContext);
-  const posthog = usePostHog();
-  const dispatch = useDispatch();
-  const [paused, setPaused] = useState<boolean | undefined>(undefined);
-  const [update, setUpdate] = useState<IndexingProgressUpdate>({
-    desc: "Loading indexing config",
-    progress: 0.0,
-    status: "loading",
-  });
-
-  useWebviewListener("indexProgress", async (data) => {
-    setUpdate(data);
-  });
-
-  useEffect(() => {
-    // Triggers retrieval for possible non-default states set prior to IndexingProgressBar initialization
-    ideMessenger.post("index/indexingProgressBarInitialized", undefined);
-  }, []);
-
-  useEffect(() => {
-    if (paused === undefined) return;
-    ideMessenger.post("index/setPaused", paused);
-  }, [paused]);
-
-  function onClickRetry() {
-    // For now, we don't show in JetBrains since the re-index command
-    // is not yet implemented
-    if (update.shouldClearIndexes && !isJetBrains()) {
-      dispatch(setShowDialog(true));
-      dispatch(
-        setDialogMessage(
-          <ConfirmationDialog
-            title="Rebuild codebase index"
-            confirmText="Rebuild"
-            text={
-              "Your index appears corrupted. We recommend clearing and rebuilding it, " +
-              "which may take time for large codebases.\n\n" +
-              "For a faster rebuild without clearing data, press 'Shift + Command + P' to open " +
-              "the Command Palette, and type out 'Continue: Force Codebase Re-Indexing'"
-            }
-            onConfirm={() => {
-              posthog.capture("rebuild_index_clicked");
-              ideMessenger.post("index/forceReIndex", {
-                shouldClearIndexes: true,
-              });
-            }}
-          />,
-        ),
-      );
-    } else {
-      ideMessenger.post("index/forceReIndex", undefined);
-    }
-  }
-
-  function onClick() {
-    switch (update.status) {
-      case "failed":
-        onClickRetry();
-        break;
-      case "indexing":
-      case "loading":
-      case "paused":
-        if (update.progress < 1 && update.progress >= 0) {
-          setPaused((prev) => !prev);
-        } else {
-          ideMessenger.post("index/forceReIndex", undefined);
-        }
-        break;
-      case "disabled":
-        ideMessenger.post("openConfigJson", undefined);
-        break;
-      case "done":
-        ideMessenger.post("index/forceReIndex", undefined);
-      default:
-        break;
-    }
-  }
-
-  return (
-    <div className="mt-6 flex flex-col">
-      <div className="mb-0 flex justify-between text-sm">
-        <IndexingProgressTitleText update={update} />
-        {update.status !== "loading" && (
-          <IndexingProgressIndicator update={update} />
-        )}
-      </div>
-
-      <IndexingProgressBar update={update} />
-
-      <IndexingProgressSubtext update={update} onClick={onClick} />
-
-      {update.status === "failed" && (
-        <div className="mt-4">
-          <IndexingProgressErrorText update={update} />
-        </div>
-      )}
-    </div>
-  );
-}
-
-function AllDocsIndexingProgress() {}
-export default AllDocsIndexingProgress;
